requirements.txt
Django==2.2.1
djangorestframework
jsonfield
